third_party/WebKit/LayoutTests/webaudio/audiocontext-max-contexts.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/compatibility.js"></script>
</head>
<body>
<script>
description("Check that the AudioContext constructor throws when the limit on hardware contexts is reached.");

function reachHardwareContextsLimit() {
    var context = [];
    for (var i=0; i<10; ++i)
        context[i] = new AudioContext();
}

shouldThrow("reachHardwareContextsLimit()");
</script>
</body>
</html>
